Remove duplicate navigation when selecting issue.
This fixes an issue where, after selecting an issue,
it would take two presses of the browser back button
to navigate back to the previous page. This issue was
uncovered (indirectly) while triaging #216.

Improved unit test coverage, but, as per discussion
in #219, we should add an e2e test to verify that
clicking on the link successfully navigates to the
issue.